ProcIP_Module2-main/backend/requirements.txt
flask==3.0.0
pandas>=2.2.0
openpyxl>=3.1.5
xlsxwriter==3.1.9
requests>=2.31.0
python-dotenv==1.0.0
openai>=1.12.0
pywebview==4.4.1
werkzeug==3.0.1
langdetect==1.0.9
flask-cors>=4.0.0
